# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Best backend ever.",
        "title": "Mementor back",
        "contact": {
            "name": "@withoutasecondthought",
            "email": "mrmarkeld@gmail.com"
        },
        "version": "1.0"
    },
    "host": "api.ilyaprojects.com",
    "basePath": "/mementor",
    "paths": {
        "/book": {
            "post": {
                "description": "book mentor of your dream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "New Booking",
                "parameters": [
                    {
                        "description": "params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/mentor": {
            "get": {
                "description": "get your page",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "Show your Page",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetMentorResponse"
                        }
                    },
                    "500": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Send full mentor info to update your profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "change mentor",
                "parameters": [
                    {
                        "description": "Account info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PutMentorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "400": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "500": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "remove mentor from bd",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "Delete mentor",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "500": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/mentor/image": {
            "post": {
                "description": "Upload your best photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "Upload Image",
                "parameters": [
                    {
                        "description": "base64",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/Image"
                        }
                    },
                    "400": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "500": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/mentor/{id}": {
            "get": {
                "description": "Give you mentor without personal fields",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "Show a mentor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GetMentorResponse"
                        }
                    },
                    "404": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/mentor/{page}": {
            "post": {
                "description": "get mentors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mentor"
                ],
                "summary": "return List of Mentors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "number of page",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "params",
                        "name": "params",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/PostMentorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PostMentorResponse"
                        }
                    },
                    "400": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "500": {
                        "description": "error occurred",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/sign-in": {
            "post": {
                "description": "sign in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "sign in",
                "operationId": "sign-in",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "userinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PostAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        },
        "/sign-up": {
            "post": {
                "description": "sign up",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "sign up",
                "operationId": "sign-up",
                "parameters": [
                    {
                        "description": "Account data",
                        "name": "userinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PostAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "BasicResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Education": {
            "type": "object",
            "required": [
                "department",
                "place"
            ],
            "properties": {
                "department": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                }
            }
        },
        "GetMentorResponse": {
            "type": "object",
            "required": [
                "email",
                "experienceSince",
                "grade",
                "name",
                "programmingLanguage",
                "surname",
                "tariff",
                "technology"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "canHelpWith": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "classesDone": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "education": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Education"
                    }
                },
                "email": {
                    "type": "string"
                },
                "experienceSince": {
                    "type": "integer"
                },
                "grade": {
                    "type": "string",
                    "enum": [
                        "junior",
                        "middle",
                        "senior"
                    ]
                },
                "image": {
                    "$ref": "#/definitions/Image"
                },
                "language": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "programmingLanguage": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "surname": {
                    "type": "string"
                },
                "tariff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Tariff"
                    }
                },
                "technology": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "validProfile": {
                    "type": "boolean"
                }
            }
        },
        "Image": {
            "type": "object",
            "properties": {
                "144x144": {
                    "type": "string"
                },
                "512x512": {
                    "type": "string"
                }
            }
        },
        "Mentor": {
            "type": "object",
            "required": [
                "grade",
                "name",
                "programmingLanguage",
                "surname",
                "tariff"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "grade": {
                    "type": "string",
                    "enum": [
                        "junior",
                        "middle",
                        "senior"
                    ]
                },
                "image": {
                    "$ref": "#/definitions/Image"
                },
                "language": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "programmingLanguage": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "surname": {
                    "type": "string"
                },
                "tariff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Tariff"
                    }
                }
            }
        },
        "PostAuthRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "PostAuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "Bearer token"
                }
            }
        },
        "PostBookingRequest": {
            "type": "object",
            "required": [
                "customerName",
                "customerTelegram",
                "mentorId"
            ],
            "properties": {
                "customerName": {
                    "type": "string"
                },
                "customerTelegram": {
                    "type": "string"
                },
                "mentorId": {
                    "type": "string"
                },
                "tariffIndex": {
                    "type": "integer",
                    "maximum": 2,
                    "minimum": 0
                }
            }
        },
        "PostImageRequest": {
            "type": "object",
            "properties": {
                "base64": {
                    "type": "string"
                }
            }
        },
        "PostMentorRequest": {
            "type": "object",
            "properties": {
                "experienceSince": {
                    "type": "integer"
                },
                "grade": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "maxPrice": {
                    "type": "integer"
                },
                "minPrice": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                }
            }
        },
        "PostMentorResponse": {
            "type": "object",
            "required": [
                "mentors",
                "pages"
            ],
            "properties": {
                "mentors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Mentor"
                    }
                },
                "pages": {
                    "type": "integer"
                }
            }
        },
        "PutMentorRequest": {
            "type": "object",
            "required": [
                "email",
                "experienceSince",
                "grade",
                "name",
                "programmingLanguage",
                "surname",
                "tariff",
                "technology"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "canHelpWith": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "classesDone": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "education": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Education"
                    }
                },
                "email": {
                    "type": "string"
                },
                "experienceSince": {
                    "type": "integer"
                },
                "grade": {
                    "type": "string",
                    "enum": [
                        "junior",
                        "middle",
                        "senior"
                    ]
                },
                "image": {
                    "$ref": "#/definitions/Image"
                },
                "language": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "programmingLanguage": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "surname": {
                    "type": "string"
                },
                "tariff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Tariff"
                    }
                },
                "technology": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "validProfile": {
                    "type": "boolean"
                }
            }
        },
        "Tariff": {
            "type": "object",
            "required": [
                "description",
                "name",
                "price"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAUth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}